+++ datafusion/core/src/physical_optimizer/mod.rs
@@ -35,7 +35,6 @@ pub mod sort_enforcement;
 mod sort_pushdown;
 mod utils;
 
-pub mod pipeline_fixer;
 #[cfg(test)]
 pub mod test_utils;
